fix(rubocop): split goods nomenclature nested set modules
Structure-only: extract ancestor/descendant/measure associations,
populators and predicates into NestedSet::* concerns without a
ClassMethods registration step. historical_children stays in
descendant associations.

Issue: No ticket/issue
